imap: Added secured and ssl_secured flags
This adds two new boolean fields: 'secured', and 'ssl_secured' to the
imap client structure, these reflect the values of conn_secured and
conn_ssl_secured respectively.

This is a prerequisite to integrating the imap_capability_list APIs into
imap because imap needs to distinguish between secured/tls/non-secured
connections when determining whether conditional capabilities should be
displayed or not.