lib-storage: mailbox_list_get_storage() - Add flags, and allow changing vname
The previous API allowed changing the mailbox_list, but not the vname
string. Now it's possible to change both.

The flags allows the implementation more flexibility. It will be used by
following commits to determine whether opening a virtual mailbox should
actually be opening the backend mailbox (for saving mails).