global: Use is_config_binary() where CONFIG_BINARY macro does not propagate
Not all the *-settings.c sources are embeded in the generated all-settings.c
source file, which is the only scope where the CONFIG_BINARY macro is defined.
This specifically applies to master-service-ssl-settings.c which contained
code never activated by the #ifdef(s) on the macro as this file is not part
of the synthesized code.