lib: unichar - Don't return 0 from uni_utf8_get_char_n() when NUL byte is encountered.
This caused problems for code that uses this function to read from a
size-limited buffer that can contain NUL bytes; i.e. lib-smtp. In that case,
hitting a NUL byte yields the same result as hitting the end of the buffered
data, which is unacceptable. No other code relies on the function returning 0,
so this can be changed safely to returning -1 instead.