m4: glibc.m4 - Improve glibc-check
The previous check was based on a compilation error and negating the
value. This updated check is more clear as it runs the preprocessor with
the appropriate header file and checks if the output matches, which in
turn allows for a more forward readable definition of the
i_cv_have_glibc value.