doveadm: Move username printing from server side to client side
The client requests output for only a single user from the server side, so
it makes sense that the client is the one printing the username.

This also changes sticky headers to be usable only for the username.
Unfortunately it's not easy to remove support for sticky header entirely,
because the username must not be printed before the first proper
doveadm_print() call. For example if "doveadm fetch" doesn't find any
mails, it shouldn't print the username either.