lib: Always evaluate e_error(), e_warning() and e_info() parameters
e_debug() still delays evaluation until it's known that the event won't
simply be discarded. Since there are lots of e_debug() calls, this is a
good performance optimization. Other types of logging are much more rare
though and it's not so useful as an optimization. It can also be
unexpected that the evaluation doesn't happen. Especially with e_error()
and e_warning() it is expected, since it's only some special circumstances
where they don't log the output.