imap: Add UTF8=ACCEPT capability (RFC 6855)
This adds the capability and accepts UTF8 quoted-strings.

The SEARCH command is changed as required by RFC 6855, section 3, final
paragraph. There are no changes to the searching, as 6855 only changes
changes the syntax.

This does not add any kind of downgrading. Before this change, Dovecot would
accept an APPEND of a message such as

   From: grå@grå.org
   Subject: grå
   ...

and would just-send-8 that to any IMAP clients. This change maintains that policy.