fts: fts_indexer_input_args() - Fix client-side timeout
Client side timeout now just triggers if there is period long enough
where the server doesn't send updates. This works just because the
server is not pushing the updates properly.

The client must timeout BOTH if the server is unresponsive AND if
the server sends updates but the overall elapsed time exceedes.