indexer: index_mailbox_precache() - Ensure updates are sent even with few messages and very slow processing
With the former behavior, if we have 101 messages and they take 30 seconds,
the client sees the progress stuck at 0% for most the processign time, only
to see it jump to 99% a fraction of a second before the competion.

This change ensures that the client sees a timely update every time the
integer percentage actually changes.