config, lib-master: Optimize binary config format - split settings blocks by their structs
This allows looking up settings only for the requested settings struct
without reading all the other settings as well.

Additionally this new format allows delaying error handling. For example
if "ssl_key = <path" handling can't open the path because it doesn't have
permissions, the error is delayed until SSL server settings are actually
used by the config client. This is beneficial when the client never even
attempts to use the SSL server settings, in which case the lack of
permission is irrelevant and no error is ever seen. This example
behavior doesn't actually yet work in this commit though.