lib-settings, config: Allow empty filter names if they are explicitly quoted
This will be required to keep allowing mailbox "" { ... } to be used for
matching the namespace prefix mailbox.

The empty name is now escaped as "\." string. It might have been possible
to make it work as an empty string, but it would have been rather confusing
(e.g. "mailbox= ", "mailbox//auto").