config: Add support for changing the default settings internally
Currently the default settings can be given only via the defaults struct.
With this change it's possible to separately track whether settings changes
are intended to be defaults or explicit configuration changes.

Only the explicit config changes are visible in e.g. "doveconf -n" output,
while all the changes will be written to the binary config file.