lib-settings, config: Add boolean-list (SET_BOOLLIST) type
The setting can also be configured as a comma/space separated list:

  list = key1 key2

It can further be modified with SET_STRLIST-like syntax:

list {
  key1 = no
  key3 = yes
  key4 = %{variable}
}

THe end result is an array of keys that were "yes". In this case key2, key3
and maybe key4 depending on the variable.

It's possible to have spaces also in keys, e.g.:

list = "key1 with space", key2
list {
  "key3 with space" = yes
}