lib-http: http_client_connection_get_settings() - Assert that peer is non-NULL
This change is only temporary. A later commit reimplements it by using
the last assigned http_client's settings.

It shouldn't be possible with current code for peer to be NULL, so this
temporary change in behavior doesn't matter. If it had been possible for
peer to be NULL, the current code would have always returned empty
settings, since http_client_context.set is never set outside unit tests.
This is equivalent to the new way of explicitly returning default settings.