doveadm fs: Change fs-driver & fs-args parameters to filter-name
The filter-name refers to the filter section in config file that contains
the fs_driver and other necessary settings. It's also possible to specify
all necessary settings using -o parameter, e.g.:

doveadm \
  -o fs_parent/fs_driver=posix \
  -o fs_driver=randomfail \
  -o fs_randomfail_ops/read=50 \
  fs get "" /path/to/file

Note that currently the order of fs_driver settings is important here,
because the first match is used. This may change in the future.