auth: auth-request - Move most of the log prefix to the main request event
This way, any request-related logging will have a prefix with all the necessary
information. Before, the main request event had no prefix and when no sub-event
was available no prefix was added to log lines, meaning that the request these
related to could not be determined. Sub-events will now have a smaller
additional prefix that relates to the subsystem alone.