lib-master: stats_client_wait() - Distinguish between handshake and deinit
stats_client_wait() is used to wait for the handshake response as well
as waiting for the stats flushing on stats_client_deinit(). Previously
the stats_client_deinit() could easily have timed out.

This adds a specific error message for the deinit timeout as well as a
bigger timeout (30 seconds).