lib-storage: shared - Remove check to see if shared namespace's storage exists
This doesn't work with imapc when local index files are used, because
mailbox_list_get_root_path() returns the index file location then. Just
because it doesn't exist yet doesn't mean that the namespace is unusable.

There is already a check to verify whether the user exists, which should be
enough.