auth: Add generic configuration for passdb/userdb fields
The following commits will make these work like:

passdb foo {
  fields {
    key = %{passdb_driver:fieldname:default}
  }
}

And similarly for userdb. This will remove the need for default_fields and
override_fields settings. It's also possible to use
passdb/userdb_fields_import_all setting to import all fields, which is
going to be especially useful for passwd-file.