auth: passdb/userdb sql - Convert to new settings
sql_drivers_deinit() needed to be moved earlier, because the sql_db is
now kept in cache until it's called. In the earlier code the sql_db was
deinitialized when the passdb/userdb sql was deinitialized. If the
sql_db deinit happens too late, it's late in aborting requests and that
causes crashes.