lib-master, config: Add a required dovecot_storage_version setting
This setting specifies specifies the oldest Dovecot version that must be
able to read the files written by this Dovecot. This setting replaces
separate settings that are added when file formats are changed.

The intention is that when upgrading Dovecot cluster, this setting is
first kept as the old Dovecot version. Once the cluster is fully upgraded
to a new version and there is no intention to rollback to the old version
anymore, this version number can be increased.