quota: Drop support for percentage based quota_grace
Also change the default from 10% to 10 MB. It's likely not necessary to
support both, and having a fixed size makes usually more sense than a
percentage that can be rather huge. If we wanted to support both, the
settings redesign would require two separate grace settings, which seems
a bit excessive.