auth: Remove empty password check for nopassword
The new fields { .. } way makes this a bit more difficult. Previously
it was ok to put nopassword to default_fields or override_fields even
if the passdb itself returned a password - this empty password check
was done only if the passdb internally returned both password and
nopassword. However, now it's all in the same fields { .. } where it's
not always easy to prevent adding the password.