lib: ostream-file - Unset TCP_NODELAY only when stream is written uncorked
The previous behavior was always to set TCP_NODELAY and unset it immediately
when uncorking the stream. But if stream is only written to while it's
corked, this adds unnecessary setsockopt() calls for every uncork.

The current behavior is to unset TCP_NODELAY only if data is written to
ostream while it's not corked.