config: Change the generated settings-history-core from .h to .c
This changes the output file to a header file instead of a C file, this helps meson deal
better with dependencies on custom_targets.

It is technically an #included header file so it makes sense.

The ideal solution would be to keep it as a C file, and add forward declarations of the
arrays defined in it instead of verbatim #include'ing it. However, array forward decls are
not possible without a specified size which cannot be provided in advance and declaring
them as pointers is perhaps not the best idea.