m4: Fix for systems with signed 32-bit time_t
The code and tests have some hard-coded values based on the size of
time_t but they don't take into account whether or not time_t is a
signed value.

On systems such as i686 with time_t as a signed 32-bit integer, the
maximum value is the same as an unsigned 31-bit integer. Tweak the
configure test to treat 32-bit signed as 31 bits.