lib-auth-client: Fix user iteration potentially listing empty usernames
Broken by 529e9d98cb637fc69bd5e8a3ed769b0031558c67

With earlier code one of ctx->username, finished or failed was guaranteed to
be set. With the new auth_master_request_wait() code this is no longer true.
It's unclear when exactly this happens or how to reproduce, but the fixed code
makes sure it doesn't, or at least assert-crashes if it still does.